Move `StatementCategoryFilter` to separate module
commit-id:df5a0f09
