+++ crates/cairo-coverage/src/input/filter/mod.rs
@@ -1,0 +1,1 @@
+pub mod statement_category_filter;
