+++ crates/cairo-coverage/src/input/mod.rs
@@ -1,9 +1,9 @@
 mod data;
+mod filter;
 mod sierra_to_cairo_map;
-mod statement_category_filter;
 mod unique_executed_sierra_ids;
 
 pub use data::InputData;
+pub use filter::statement_category_filter::StatementCategoryFilter;
 pub use sierra_to_cairo_map::{create_sierra_to_cairo_map, SierraToCairoMap};
-pub use statement_category_filter::StatementCategoryFilter;
 pub use unique_executed_sierra_ids::UniqueExecutedSierraIds;
